+++ zenjxl-decoder/src/render/stages/blending.rs
@@ -163,7 +163,7 @@ mod test {
     #[test]
     fn blending_consistency() -> Result<()> {
         let (file_header, frame_header, _) =
-            read_headers_and_toc(include_bytes!("../../../resources/test/basic.jxl")).unwrap();
+            read_headers_and_toc(&crate::util::test::fixture_bytes("basic.jxl")).unwrap();
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(0);
         let reference_frames = Arc::new([
             Some(ReferenceFrame::random(&mut rng, 500, 500, 4, false)?),