+++ zenjxl-decoder/src/render/stages/patches.rs
@@ -73,7 +73,7 @@ mod test {
     #[test]
     fn patches_consistency() -> Result<()> {
         let (file_header, _, _) =
-            read_headers_and_toc(include_bytes!("../../../resources/test/basic.jxl")).unwrap();
+            read_headers_and_toc(&crate::util::test::fixture_bytes("basic.jxl")).unwrap();
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(0);
         let patch_dict = Arc::new(PatchesDictionary::random(
             (500, 500),